move introduction into own js file
